+++ modules/AWSPowerShell/Cmdlets/Connect/AWS.Tools.Connect.Completers.psm1
@@ -133,6 +133,16 @@ $CONN_Completers = {
             break
         }
 
+        # Amazon.Connect.EventSourceName
+        {
+            ($_ -eq "Get-CONNRuleList/EventSourceName") -Or
+            ($_ -eq "New-CONNRule/TriggerEventSource_EventSourceName")
+        }
+        {
+            $v = "OnPostCallAnalysisAvailable","OnPostChatAnalysisAvailable","OnRealTimeCallAnalysisAvailable","OnSalesforceCaseCreate","OnZendeskTicketCreate","OnZendeskTicketStatusUpdate"
+            break
+        }
+
         # Amazon.Connect.HierarchyGroupMatchType
         "Search-CONNUser/SearchCriteria_HierarchyGroupCondition_HierarchyGroupMatchType"
         {
@@ -208,6 +218,17 @@ $CONN_Completers = {
         }
         {
             $v = "PHONE_NUMBER","QUEUE","USER"
+            break
+        }
+
+        # Amazon.Connect.RulePublishStatus
+        {
+            ($_ -eq "Get-CONNRuleList/PublishStatus") -Or
+            ($_ -eq "New-CONNRule/PublishStatus") -Or
+            ($_ -eq "Update-CONNRule/PublishStatus")
+        }
+        {
+            $v = "DRAFT","PUBLISHED"
             break
         }
 
@@ -310,12 +331,14 @@ $CONN_map = @{
     "AttributeType"=@("Get-CONNInstanceAttribute","Update-CONNInstanceAttribute")
     "ContactFlowModuleState"=@("Get-CONNContactFlowModuleList")
     "ContactFlowState"=@("Update-CONNContactFlowMetadata")
+    "EventSourceName"=@("Get-CONNRuleList")
     "IdentityManagementType"=@("New-CONNInstance")
     "IntegrationType"=@("Get-CONNIntegrationAssociationList","New-CONNIntegrationAssociation")
     "LanguageCode"=@("Add-CONNDefaultVocabulary","Get-CONNDefaultVocabularyList","New-CONNVocabulary","Search-CONNVocabulary")
     "LexVersion"=@("Get-CONNBotList")
     "PhoneNumberCountryCode"=@("Search-CONNAvailablePhoneNumber")
     "PhoneNumberType"=@("Search-CONNAvailablePhoneNumber")
+    "PublishStatus"=@("Get-CONNRuleList","New-CONNRule","Update-CONNRule")
     "QuickConnectConfig_QuickConnectType"=@("New-CONNQuickConnect","Update-CONNQuickConnectConfig")
     "ResourceType"=@("Add-CONNInstanceStorageConfig","Get-CONNInstanceStorageConfig","Get-CONNInstanceStorageConfigList","Remove-CONNInstanceStorageConfig","Update-CONNInstanceStorageConfig")
     "SearchCriteria_HierarchyGroupCondition_HierarchyGroupMatchType"=@("Search-CONNUser")
@@ -328,6 +351,7 @@ $CONN_map = @{
     "StorageConfig_S3Config_EncryptionConfig_EncryptionType"=@("Add-CONNInstanceStorageConfig","Update-CONNInstanceStorageConfig")
     "StorageConfig_StorageType"=@("Add-CONNInstanceStorageConfig","Update-CONNInstanceStorageConfig")
     "TrafficType"=@("Start-CONNOutboundVoiceContact")
+    "TriggerEventSource_EventSourceName"=@("New-CONNRule")
     "Type"=@("New-CONNContactFlow")
     "UseCaseType"=@("New-CONNUseCase")
     "VoiceRecordingConfiguration_VoiceRecordingTrack"=@("Start-CONNContactRecording")
@@ -403,6 +427,7 @@ $CONN_SelectMap = @{
                "New-CONNQueue",
                "New-CONNQuickConnect",
                "New-CONNRoutingProfile",
+               "New-CONNRule",
                "New-CONNSecurityProfile",
                "New-CONNTaskTemplate",
                "New-CONNTrafficDistributionGroup",
@@ -416,6 +441,7 @@ $CONN_SelectMap = @{
                "Remove-CONNInstance",
                "Remove-CONNIntegrationAssociation",
                "Remove-CONNQuickConnect",
+               "Remove-CONNRule",
                "Remove-CONNSecurityProfile",
                "Remove-CONNTaskTemplate",
                "Remove-CONNTrafficDistributionGroup",
@@ -435,6 +461,7 @@ $CONN_SelectMap = @{
                "Get-CONNQueue",
                "Get-CONNQuickConnect",
                "Get-CONNRoutingProfile",
+               "Get-CONNRule",
                "Get-CONNSecurityProfile",
                "Get-CONNTrafficDistributionGroup",
                "Get-CONNUser",
@@ -480,6 +507,7 @@ $CONN_SelectMap = @{
                "Get-CONNQuickConnectList",
                "Get-CONNRoutingProfileQueueList",
                "Get-CONNRoutingProfileList",
+               "Get-CONNRuleList",
                "Get-CONNSecurityKeyList",
                "Get-CONNSecurityProfilePermissionList",
                "Get-CONNSecurityProfileList",
@@ -536,6 +564,7 @@ $CONN_SelectMap = @{
                "Update-CONNRoutingProfileDefaultOutboundQueue",
                "Update-CONNRoutingProfileName",
                "Update-CONNRoutingProfileQueue",
+               "Update-CONNRule",
                "Update-CONNSecurityProfile",
                "Update-CONNTaskTemplate",
                "Update-CONNTrafficDistribution",