+++ modules/AWSPowerShell/Cmdlets/Connect/Basic/New-CONNRule-Cmdlet.cs
@@ -1,0 +1,420 @@
+/*******************************************************************************
+ *  Copyright 2012-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.Connect;
+using Amazon.Connect.Model;
+
+namespace Amazon.PowerShell.Cmdlets.CONN
+{
+    /// <summary>
+    /// Creates a rule for the specified Amazon Connect instance.
+    /// </summary>
+    [Cmdlet("New", "CONNRule", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
+    [OutputType("Amazon.Connect.Model.CreateRuleResponse")]
+    [AWSCmdlet("Calls the Amazon Connect Service CreateRule API operation.", Operation = new[] {"CreateRule"}, SelectReturnType = typeof(Amazon.Connect.Model.CreateRuleResponse))]
+    [AWSCmdletOutput("Amazon.Connect.Model.CreateRuleResponse",
+        "This cmdlet returns an Amazon.Connect.Model.CreateRuleResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class NewCONNRuleCmdlet : AmazonConnectClientCmdlet, IExecutor
+    {
+        
+        #region Parameter Action
+        /// <summary>
+        /// <para>
+        /// <para>A list of actions to be run when the rule is triggered.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyCollection]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        [Alias("Actions")]
+        public Amazon.Connect.Model.RuleAction[] Action { get; set; }
+        #endregion
+        
+        #region Parameter TriggerEventSource_EventSourceName
+        /// <summary>
+        /// <para>
+        /// <para>The name of the event source.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        [AWSConstantClassSource("Amazon.Connect.EventSourceName")]
+        public Amazon.Connect.EventSourceName TriggerEventSource_EventSourceName { get; set; }
+        #endregion
+        
+        #region Parameter Function
+        /// <summary>
+        /// <para>
+        /// <para>The conditions of the rule.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String Function { get; set; }
+        #endregion
+        
+        #region Parameter InstanceId
+        /// <summary>
+        /// <para>
+        /// <para>The identifier of the Amazon Connect instance. You can find the instanceId in the
+        /// ARN of the instance.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String InstanceId { get; set; }
+        #endregion
+        
+        #region Parameter TriggerEventSource_IntegrationAssociationId
+        /// <summary>
+        /// <para>
+        /// <para>The identifier for the integration association.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String TriggerEventSource_IntegrationAssociationId { get; set; }
+        #endregion
+        
+        #region Parameter Name
+        /// <summary>
+        /// <para>
+        /// <para>A unique name for the rule.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true)]
+        #else
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String Name { get; set; }
+        #endregion
+        
+        #region Parameter PublishStatus
+        /// <summary>
+        /// <para>
+        /// <para>The publish status of the rule.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        [AWSConstantClassSource("Amazon.Connect.RulePublishStatus")]
+        public Amazon.Connect.RulePublishStatus PublishStatus { get; set; }
+        #endregion
+        
+        #region Parameter ClientToken
+        /// <summary>
+        /// <para>
+        /// <para>A unique, case-sensitive identifier that you provide to ensure the idempotency of
+        /// the request. If not provided, the Amazon Web Services SDK populates this field. For
+        /// more information about idempotency, see <a href="https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/">Making
+        /// retries safe with idempotent APIs</a>.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String ClientToken { get; set; }
+        #endregion
+        
+        #region Parameter Select
+        /// <summary>
+        /// Use the -Select parameter to control the cmdlet output. The default value is '*'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.Connect.Model.CreateRuleResponse).
+        /// Specifying the name of a property of type Amazon.Connect.Model.CreateRuleResponse will result in that property being returned.
+        /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public string Select { get; set; } = "*";
+        #endregion
+        
+        #region Parameter PassThru
+        /// <summary>
+        /// Changes the cmdlet behavior to return the value passed to the Name parameter.
+        /// The -PassThru parameter is deprecated, use -Select '^Name' instead. This parameter will be removed in a future version.
+        /// </summary>
+        [System.Obsolete("The -PassThru parameter is deprecated, use -Select '^Name' instead. This parameter will be removed in a future version.")]
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public SwitchParameter PassThru { get; set; }
+        #endregion
+        
+        #region Parameter Force
+        /// <summary>
+        /// This parameter overrides confirmation prompts to force 
+        /// the cmdlet to continue its operation. This parameter should always
+        /// be used with caution.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public SwitchParameter Force { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var resourceIdentifiersText = FormatParameterValuesForConfirmationMsg(nameof(this.Name), MyInvocation.BoundParameters);
+            if (!ConfirmShouldProceed(this.Force.IsPresent, resourceIdentifiersText, "New-CONNRule (CreateRule)"))
+            {
+                return;
+            }
+            
+            var context = new CmdletContext();
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            #pragma warning disable CS0618, CS0612 //A class member was marked with the Obsolete attribute
+            if (ParameterWasBound(nameof(this.Select)))
+            {
+                context.Select = CreateSelectDelegate<Amazon.Connect.Model.CreateRuleResponse, NewCONNRuleCmdlet>(Select) ??
+                    throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
+                if (this.PassThru.IsPresent)
+                {
+                    throw new System.ArgumentException("-PassThru cannot be used when -Select is specified.", nameof(this.Select));
+                }
+            }
+            else if (this.PassThru.IsPresent)
+            {
+                context.Select = (response, cmdlet) => this.Name;
+            }
+            #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
+            if (this.Action != null)
+            {
+                context.Action = new List<Amazon.Connect.Model.RuleAction>(this.Action);
+            }
+            #if MODULAR
+            if (this.Action == null && ParameterWasBound(nameof(this.Action)))
+            {
+                WriteWarning("You are passing $null as a value for parameter Action which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.ClientToken = this.ClientToken;
+            context.Function = this.Function;
+            #if MODULAR
+            if (this.Function == null && ParameterWasBound(nameof(this.Function)))
+            {
+                WriteWarning("You are passing $null as a value for parameter Function which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.InstanceId = this.InstanceId;
+            #if MODULAR
+            if (this.InstanceId == null && ParameterWasBound(nameof(this.InstanceId)))
+            {
+                WriteWarning("You are passing $null as a value for parameter InstanceId which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.Name = this.Name;
+            #if MODULAR
+            if (this.Name == null && ParameterWasBound(nameof(this.Name)))
+            {
+                WriteWarning("You are passing $null as a value for parameter Name which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.PublishStatus = this.PublishStatus;
+            #if MODULAR
+            if (this.PublishStatus == null && ParameterWasBound(nameof(this.PublishStatus)))
+            {
+                WriteWarning("You are passing $null as a value for parameter PublishStatus which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.TriggerEventSource_EventSourceName = this.TriggerEventSource_EventSourceName;
+            #if MODULAR
+            if (this.TriggerEventSource_EventSourceName == null && ParameterWasBound(nameof(this.TriggerEventSource_EventSourceName)))
+            {
+                WriteWarning("You are passing $null as a value for parameter TriggerEventSource_EventSourceName which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.TriggerEventSource_IntegrationAssociationId = this.TriggerEventSource_IntegrationAssociationId;
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.Connect.Model.CreateRuleRequest();
+            
+            if (cmdletContext.Action != null)
+            {
+                request.Actions = cmdletContext.Action;
+            }
+            if (cmdletContext.ClientToken != null)
+            {
+                request.ClientToken = cmdletContext.ClientToken;
+            }
+            if (cmdletContext.Function != null)
+            {
+                request.Function = cmdletContext.Function;
+            }
+            if (cmdletContext.InstanceId != null)
+            {
+                request.InstanceId = cmdletContext.InstanceId;
+            }
+            if (cmdletContext.Name != null)
+            {
+                request.Name = cmdletContext.Name;
+            }
+            if (cmdletContext.PublishStatus != null)
+            {
+                request.PublishStatus = cmdletContext.PublishStatus;
+            }
+            
+             // populate TriggerEventSource
+            var requestTriggerEventSourceIsNull = true;
+            request.TriggerEventSource = new Amazon.Connect.Model.RuleTriggerEventSource();
+            Amazon.Connect.EventSourceName requestTriggerEventSource_triggerEventSource_EventSourceName = null;
+            if (cmdletContext.TriggerEventSource_EventSourceName != null)
+            {
+                requestTriggerEventSource_triggerEventSource_EventSourceName = cmdletContext.TriggerEventSource_EventSourceName;
+            }
+            if (requestTriggerEventSource_triggerEventSource_EventSourceName != null)
+            {
+                request.TriggerEventSource.EventSourceName = requestTriggerEventSource_triggerEventSource_EventSourceName;
+                requestTriggerEventSourceIsNull = false;
+            }
+            System.String requestTriggerEventSource_triggerEventSource_IntegrationAssociationId = null;
+            if (cmdletContext.TriggerEventSource_IntegrationAssociationId != null)
+            {
+                requestTriggerEventSource_triggerEventSource_IntegrationAssociationId = cmdletContext.TriggerEventSource_IntegrationAssociationId;
+            }
+            if (requestTriggerEventSource_triggerEventSource_IntegrationAssociationId != null)
+            {
+                request.TriggerEventSource.IntegrationAssociationId = requestTriggerEventSource_triggerEventSource_IntegrationAssociationId;
+                requestTriggerEventSourceIsNull = false;
+            }
+             // determine if request.TriggerEventSource should be set to null
+            if (requestTriggerEventSourceIsNull)
+            {
+                request.TriggerEventSource = null;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                object pipelineOutput = null;
+                pipelineOutput = cmdletContext.Select(response, this);
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.Connect.Model.CreateRuleResponse CallAWSServiceOperation(IAmazonConnect client, Amazon.Connect.Model.CreateRuleRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon Connect Service", "CreateRule");
+            try
+            {
+                #if DESKTOP
+                return client.CreateRule(request);
+                #elif CORECLR
+                return client.CreateRuleAsync(request).GetAwaiter().GetResult();
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public List<Amazon.Connect.Model.RuleAction> Action { get; set; }
+            public System.String ClientToken { get; set; }
+            public System.String Function { get; set; }
+            public System.String InstanceId { get; set; }
+            public System.String Name { get; set; }
+            public Amazon.Connect.RulePublishStatus PublishStatus { get; set; }
+            public Amazon.Connect.EventSourceName TriggerEventSource_EventSourceName { get; set; }
+            public System.String TriggerEventSource_IntegrationAssociationId { get; set; }
+            public System.Func<Amazon.Connect.Model.CreateRuleResponse, NewCONNRuleCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response;
+        }
+        
+    }
+}